keep shared stuff in peer.go to stay consistent with storj/docs
